requirements.txt
annotated-types==0.7.0
anyio==4.8.0
arxiv==2.1.3
arxivpy==0.1.dev0
asgiref==3.8.1
attrs==24.3.0
autoflake==2.3.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.2
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
deepsearch-glm==1.0.0
Deprecated==1.2.15
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docling==2.15.0
docling-core==2.13.1
docling-ibm-models==3.1.0
docling-parse==3.0.0
durationpy==0.9
easyocr==1.7.2
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastapi-cli==0.0.7
feedparser==6.0.11
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
fonttools==4.55.3
fsspec==2024.12.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
jellyfish==1.1.3
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaleido==0.2.1
keybert==0.8.5
kiwisolver==1.4.8
kubernetes==31.0.0
lazy_loader==0.4
llama_cpp_python==0.3.5
lxml==5.3.0
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
milvus-lite==2.4.11
mmh3==5.0.1
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.17
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.5
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
plotly==5.24.1
pluggy==1.5.0
portalocker==2.10.1
posthog==3.7.5
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pymilvus==2.5.3
pyparsing==3.2.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.4
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.12.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
safetensors==0.5.2
scikit-image==0.25.0
scikit-learn==1.6.1
scipy==1.15.0
segtok==1.5.11
semchunk==2.2.2
sentence-transformers==3.3.1
setuptools==75.8.0
sgmllib3k==1.0.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sse-starlette==2.2.1
starlette==0.41.3
starlette-context==0.3.6
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tokenizers==0.21.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
unicorn==2.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.1
XlsxWriter==3.2.0
yake==0.4.8
zipp==3.21.0